Add archivematica.search package
This reorganizes and copies the functions from the
elasticSearchFunctions module into a new package, aiming to preserve
the existing functionality without changes.
