+++ src/archivematica/dashboard/main/management/commands/update_elasticsearch_mappings.py
@@ -15,7 +15,8 @@ import sys
 from django.conf import settings
 from elasticsearch import ElasticsearchException
 
-from archivematica.archivematicaCommon import elasticSearchFunctions as es
+import archivematica.search.client
+import archivematica.search.constants
 from archivematica.dashboard.main.management.commands import DashboardCommand
 
 
@@ -24,7 +25,7 @@ class Command(DashboardCommand):
 
     def handle(self, *args, **options):
         # Check that the AIPs index is enabled before proceeding.
-        if es.AIPS_INDEX not in settings.SEARCH_ENABLED:
+        if archivematica.search.constants.AIPS_INDEX not in settings.SEARCH_ENABLED:
             self.error(
                 "The AIPs indexes are not enabled. Please, make sure to "
                 "set the *_SEARCH_ENABLED environment variables to `true` "
@@ -34,34 +35,40 @@ class Command(DashboardCommand):
             sys.exit(1)
 
         try:
-            es.setup_reading_from_conf(settings)
-            es_client = es.get_client()
+            archivematica.search.client.setup_reading_from_conf(settings)
+            es_client = archivematica.search.client.get_client()
         except ElasticsearchException:
             self.error("Error: Elasticsearch may not be running.")
             sys.exit(1)
 
         # Update the AIPs index mappings.
         es_client.indices.put_mapping(
-            index=es.AIPS_INDEX,
-            doc_type=es.DOC_TYPE,
+            index=archivematica.search.constants.AIPS_INDEX,
+            doc_type=archivematica.search.constants.DOC_TYPE,
             body={
                 "properties": {
-                    es.ES_FIELD_ACCESSION_IDS: {"type": "keyword"},
-                    es.ES_FIELD_STATUS: {"type": "keyword"},
-                    es.ES_FIELD_FILECOUNT: {"type": "integer"},
-                    es.ES_FIELD_LOCATION: {"type": "keyword"},
+                    archivematica.search.constants.ES_FIELD_ACCESSION_IDS: {
+                        "type": "keyword"
+                    },
+                    archivematica.search.constants.ES_FIELD_STATUS: {"type": "keyword"},
+                    archivematica.search.constants.ES_FIELD_FILECOUNT: {
+                        "type": "integer"
+                    },
+                    archivematica.search.constants.ES_FIELD_LOCATION: {
+                        "type": "keyword"
+                    },
                 }
             },
         )
 
         # Update the AIP files index mapping.
         es_client.indices.put_mapping(
-            index=es.AIP_FILES_INDEX,
-            doc_type=es.DOC_TYPE,
+            index=archivematica.search.constants.AIP_FILES_INDEX,
+            doc_type=archivematica.search.constants.DOC_TYPE,
             body={
                 "properties": {
                     "accessionid": {"type": "keyword"},
-                    es.ES_FIELD_STATUS: {"type": "keyword"},
+                    archivematica.search.constants.ES_FIELD_STATUS: {"type": "keyword"},
                     "filePath": {
                         "type": "text",
                         "analyzer": "file_path_and_name",
@@ -74,4 +81,4 @@ class Command(DashboardCommand):
         # Perform an update by query on the aipfiles index to populate
         # the filePath.raw subfield from existing text values. We do
         # not specify a query to ensure that all documents are updated.
-        es_client.update_by_query(es.AIP_FILES_INDEX)
+        es_client.update_by_query(archivematica.search.constants.AIP_FILES_INDEX)
