+++ src/archivematica/dashboard/wsgi.py
@@ -3,9 +3,9 @@ from django.conf import settings
 from django.core.wsgi import get_wsgi_application
 
 django.setup()
-from archivematica.archivematicaCommon import elasticSearchFunctions
+from archivematica.search import client
 
 application = get_wsgi_application()
 
 # Set up Elasticsearch client
-elasticSearchFunctions.setup_reading_from_conf(settings)
+client.setup_reading_from_conf(settings)
