+++ src/archivematica/search/exceptions.py
@@ -1,0 +1,12 @@
+class SearchEngineError(Exception):
+    """Not operational errors."""
+
+    pass
+
+
+class EmptySearchResultError(SearchEngineError):
+    pass
+
+
+class TooManyResultsError(SearchEngineError):
+    pass
